+++ cli/internal/cmd/details.go
@@ -1,0 +1,7 @@
+package cmd
+
+type clusterIDFile struct {
+	ClusterID string
+	OwnerID   string
+	Endpoint  string
+}